Conflicts:
	ide/dictionary.txt
	include/StandardTimer.h
	src/time/StandardTime.cpp
	src/time/StandardTimer.cpp

+++ include/AbstractARAClient.h
@@ -215,7 +215,7 @@ protected:
     unsigned int packetDeliveryDelayInMilliSeconds;
     unsigned int routeDiscoveryTimeoutInMilliSeconds;
     unsigned int neighborActivityCheckIntervalInMilliSeconds;
-    unsigned int maxNeighborInactivityTimeInMilliSeconds;
+    long maxNeighborInactivityTimeInMilliSeconds;
     unsigned int pantIntervalInMilliSeconds;
     bool isPreviousHopFeatureActivated;
     int maxNrOfRouteDiscoveryRetries;
